Populate and style My Reviews table
