+++ app/serializers/review_serializer.rb
@@ -1,5 +1,5 @@
 class ReviewSerializer < ActiveModel::Serializer
-  attributes :id, :content, :stability, :date_visited
+  attributes :id, :content, :stability, :aesthetics, :safety, :date_visited
   belongs_to :location
   # belongs_to :user
 end
